Fix build fail on missing clock_gettime() on Wheezy (at least) 7.2 32-bit, and probably openSuse 12.2 and others.
Note: this is not using configure/automake or cmake or similar so I am unsure of the 'best' way to take into account distro idiosyncracies in this case. Tested on Wheezy 32-bit, and I know from recent experience after upgrading 11.4 --> 12.2 that -lrt is needed on Opensuse (for at least 64bit) as well.