Wean off main vs off-main lists.
Unlike bitcoin, we have a more nuanced classification of chains than
"main" and "other".  We care about the longest chain because that's
what we want to ask our peers about.  But we can't mine on that chain
until we know all the transactions, so our mining head may be a different
chain altogether.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>