Track longest descendent of known chain.
This may not be the same as the longest chain, but if not, it implies that
we're having trouble finding out the contents of blocks on the longest chain;
either we'll get those (by asking other nodes), or our known chain will
become the longest.

Thus, we should *ask* about the longest chain which exists, but tell others
about the longest chain we know most about.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>