More careful checking of batches.
We used not to check the transactions themselves; now we're going to
add backrefs for the inputs, we can properly report any errors in
a transaction, so we do that.

In addition, after checking the transactions are well-formed, we
check the merkle before other checks.  The difference is that if
the merkle matches, then any errors are actually errors in the block
which (in the next patch) we will use to taint the block and tell
all our peers about.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>