Generate messages about bad blocks.
When we discover a block contains an invalid transaction, we need to
taint it, and tell everyone.  Those messages contain enough information
to prove the problem (we never assume any peer has full knowledge).

Obviously a block which contains an invalid transaction is invalid:
we neither aim to find out more about it nor mine on any block below it.
We do remember the block and the proof: if anyone mentions it (or
an descendent) we send them the complaint message.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>