Add backrefs to transaction inputs in blocks.
This allows a node to prove that an input is invalid, by showing
the transaction referred to by the input reference.  Otherwise
a node with partial knowledge has no way of knowing if a transaction
simply references a transaction it doesn't know about (OK) or one
which doesn't exist (not OK).

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>