Major rework: group transactions explicitly by shards, not batches.
We currently group transactions into 256-tx batches; each batch is
merkled together into a top-level hash.

Now we have up to 255 transactions in each shard, and each shard is
merkled into a top-level hash.  In fact, the term "batch" goes away.

So, to begin with, there are 4 transaction hashes in each block header
(they're there even if a shard is empty).  When the network grows
this will double, etc.  There's also an 8 bit counter for the number
of transactions in each shard.

This should make network transmission simpler: you know that you need
the transactions in shard N of a batch if you're interested in an
address in N.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>