Shrink struct transaction_shard (part 1).
We're going to have a lot of these (4 per block, to start with), so
some optimization make sense: with 5 second block times we would have
over 500,000 of these within a 30-day horizon.

Since tx on the wire are always followed by the input_refs (for normal
txs) we don't need to have separate refs pointers.  We need to make
such linear structures for the case where we actually generate a block
ourselves though.

Before:
	sizeof(struct transaction_shard) = 4104
After:
	sizeof(struct transaction_shard) = 2056

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>