Allow transaction hashes in struct transaction_shard.
The PROTOCOL_PKT_SHARD packet will give us tx hashes, not the
actual transactions, so we need to keep them until we get the actual
txs.

These hashes are still useful: in particular, they are sufficient
for merkle proofs.

We could save a few bytes by making the bitmap variable length,
but it's only 32 bytes so let's leave that for now.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>