Split up check_tx() into check_tx() and check_tx_inputs().
This makes the code and error handling cleaner, and gets rid of many
private error codes.  It's a little less efficient, though, since we
need to re-lookup inputs to generate the error packet.

We also rename complain_about_input and complain_about_inputs because
the term "complain" means a problem with a block, rather than just
our feedback on a transaction.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>