Make shard more independent of block.
'struct block_shard' was not a first-class citizen: we often needed to
hand a struct block and a shard index.  That's for two reasons:
1) It didn't know its own size,
2) It could be NULL (ie. block->shard[shardnum] could be NULL).

The result was confusion when a shard wasn't yet placed in a block,
and a function referred to block->shard[shardnum] internally.

So now we always allocate it, and it contains its own size.  We don't
have to pass block around as much now.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>