Rewrite complaints.
First we break them into separate functions for the different types
of complaints.

Then we fix them: they used tal_packet_append_proof() which implicitly
generated a proof for what was currently in the block.  But bad
transactions are not supposed to be in the block, so that would fail.

1) tal_packet_append_proof() now takes an explicit proof, so hopefully
   the caller will think about what proof to generate.
2) each complaint takes a proof pointer, as it can't generate
   a proof for a transaction which isn't in the block.
3) put_tx_in_shard() doesn't generate a complaint, add check_tx_ordering()
   and make the caller do it.
4) Remove almost all the now-unused private ecodes.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>