peer: fix ownership of connection.
pettycoin: packet_io.c:22: add_log_for_fd: Assertion `fd_to_log[fd] == ((void *)0)' failed.

This is because the connection doesn't own the peer, so it doesn't get
freed when the conn is.  The destructor (destroy_peer) is responsible for
removing the fd from packet_io.c.

Since we moved to the new ccan/io (81294e06bd31c09d03f7e6870846c08055f3120a)
we want to keep the peer connected to the conn: a nice side-effect of the
simplified io_duplex.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>