Store unknown blocks.
Fascinating bug when the two seed nodes diverged by several thousand blocks
(due to I/O loss, judging from the logs).

When a new block was generated by the remote node, it sends it to us.
We don't know the ->prev_block, so we forget the block, and ask for ->prev_block.
It sends us that block, but again, we don't know prev...

While we're chasing back several thousand blocks, it generates a new block,
and we start chasing prevs for that as well.  Which slows us down more.

Resolving the FIXME and remembering detached blocks should fix this.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>