peer: initiate new block query during sync.
We currently get a surprise next time a new block is generated, and fall
back to chasing prevs.  We should fix the sync protocol, but meanwhile, this
means we at least get those blocks before we consider ourselves "synced"

Note that this breaks our sync test: we don't get told about new blocks while
we're syncing (and we do that for much longer now).  We would normally
catch up when the next block arrived.  But we don't need the block inject
at all anymore to force synchronization, so remove it.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>