pettycoin-generate: quit as soon as stdin closes.
I noticed that I was getting:

	pettycoin-generate: ''I'm not trying to cause a b-big s-s-sensation'': Broken pipe

Which can happen if pettycoin-generate is supposed to close *just* as
it is writing the solution, but should be extremely rare.

In practice, it was ignoring the zero reads from stdin, which it shouldn't;
it should ignore EWOULDBLOCK.  I renamed the function and didn't use it
where unnecessary, for increased clarity.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>