protocol: add multiple previous block hashes to block hash (BREAKS PROTOCOL)
Include the hashes of multiple previous blocks, not just the height-1
block.

The idea of such a structure (sometimes called an SPV skipchain) came
from bitcoin discussions, but it's particularly powerful for coins
with a horizon like pettycoin.

It allows discarding of some most blocks while maintaining proof that
work was done.  With a 10 second block time, we generate 3,153,600
blocks a year: 3.2GB with 1k blocks.  Using the skipchain we can weed
this down by a factor of 10 (more if we're lucky, a bit less if we're
unlucky).

I came across this idea in Mark Friedenbach's post on Compact SPV proofs:
  http://sourceforge.net/p/bitcoin/mailman/message/32111357/

But he traces the original idea to Andrew Miller's "The High-Value-Hash
Highway":
  https://bitcointalk.org/index.php?topic=98986.0

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>