recv_block: handle blocks which aren't inside protocol_pkt_block.
We're going to change protocol_pkt_welcome to include a marshaled block,
but all the code assumes a protocol_pkt_block.  Remove this assumption,
though we still need to pass in the tal'ed ptr, so we can steal it.

This lets us fix a wart in the submitblock JSON API, too: it doesn't
need the protocol_pkt_block header any more.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>