pending: Remove incorrect comment, document weird behaviour.
I noticed that after a json sendrawtransaction injection, the node
was doing a PROTOCOL_PKT_GET_TX for transactions it already had.  This is
because add_pending_tx() returns ECODE_INPUT_UNKNOWN as it doesn't look
in the pending queue for inputs.

It's a wart, but it's not a fatal flaw.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>