todo: wake peer when we complete a todo.
I noticed that my peer got stuck syncing.  It had a huge TODO list, but it
wasn't sending them out.  This is because we hit MAX_REQUESTS and then don't
wake up again when we process a request.  Fix this.

We still have a problem where our todo list gets backlogged with requests
for the children of every single block, and takes quite a while to clear.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>