welcome: duplicate the welcome packet.
If all goes well, the welcome packet gets owned by "state", which
outlives the peer.  But if it becomes a detached packet, it is owned
by that detached_block structure, and then is later discovered to be
bogus, we'll free it before peer.  In that case we'll queue an error
packet, but a correctly timed input packet could cause us to dereference
peer->welcome before that closes the connection.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>